Added level loading screen
Created a MasterLevelLoadingMenu class and scene, with added logic to the Master MainMenu. It borrows from the Continue game logic as well as the SceneLister logic.

Levels are propagated into an ItemList, with their file names cleaned. Whatever Level is activated, will update the GameStates.current_level to match the level selected. A signal emits which triggers the load action from the main menu.

It tried to keep it minimal, and imitate the style of logic already established.